refactor code, add view list view

+++ src/main/java/hotel/db/dto/checking/CheckInRequestDto.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Data
 public class CheckInRequestDto {
-    @NotNull private Long roomId;
+    @NotNull private Integer roomId;
     @NotBlank private String customerName;
 
     private String customerPhone;
